requirements.txt
boto3==1.26.90
Faker==18.3.1
python-dotenv==1.0.0